requirements.txt
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pillow>=10.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
matplotlib>=3.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

